refactor domain::Settlement
